Move receiveFee
